More splitting out and refactoring, started improving performance for getting chatList entries
