+++ src/main/deltachat/index.js
@@ -1,10 +1,7 @@
 const DeltaChat = require('deltachat-node')
 const C = require('deltachat-node/constants')
 const EventEmitter = require('events').EventEmitter
-const path = require('path')
 const log = require('../../logger').getLogger('main/deltachat')
-
-const PAGE_SIZE = 20
 
 /**
  * The Controller is the container for a deltachat instance
@@ -22,41 +19,20 @@ class DeltaChatController extends EventEmitter {
     this.loadSplitOuts()
   }
 
-  loadSplitOuts() {
+  loadSplitOuts () {
     require('./login').bind(this)()
+    require('./chatlist').bind(this)()
+    require('./messagelist').bind(this)()
+    require('./settings').bind(this)()
   }
 
-  emit(event, ...args) {
+  emit (event, ...args) {
     super.emit('ALL', event, ...args)
     super.emit(event, ...args)
   }
 
-  // Save settings for RC
-  updateSettings (saved) {
-    this._saved = saved
-  }
-
-  getPath (addr) {
-    return path.join(this.cwd, Buffer.from(addr).toString('hex'))
-  }
-
   logCoreEvent (event, payload) {
     log.debug('Core Event', event, payload)
-  }
-
-  logout () {
-    this.close()
-    this._resetState()
-
-    log.info('Logged out')
-    this.emit('logout')
-    if (typeof this._render === 'function') this._render()
-  }
-
-  close () {
-    if (!this._dc) return
-    this._dc.close()
-    this._dc = null
   }
 
   getInfo () {
@@ -67,26 +43,6 @@ class DeltaChatController extends EventEmitter {
     }
   }
 
-  sendMessage (chatId, text, filename, opts) {
-    const viewType = filename ? C.DC_MSG_FILE : C.DC_MSG_TEXT
-    const msg = this._dc.messageNew(viewType)
-    if (filename) msg.setFile(filename)
-    if (text) msg.setText(text)
-    this._dc.sendMessage(chatId, msg)
-  }
-
-  /**
-   * Update query for rendering chats with search input
-   */
-  searchChats (query) {
-    this._query = query
-    this._render()
-  }
-
-  deleteMessage (id) {
-    log.info(`deleting message ${id}`)
-    this._dc.deleteMessages(id)
-  }
 
   initiateKeyTransfer (cb) {
     return this._dc.initiateKeyTransfer(cb)
@@ -109,20 +65,6 @@ class DeltaChatController extends EventEmitter {
     log.info(`Added contact ${name} (${address})`)
     const chatId = this._dc.createChatByMessageId(deadDrop.id)
     if (chatId) this.selectChat(chatId)
-  }
-
-  unblockContact (contactId) {
-    const contact = this._dc.getContact(contactId)
-    this._dc.blockContact(contactId, false)
-    const name = contact.getNameAndAddress()
-    log.info(`Unblocked contact ${name} (id = ${contactId})`)
-  }
-
-  blockContact (contactId) {
-    const contact = this._dc.getContact(contactId)
-    this._dc.blockContact(contactId, true)
-    const name = contact.getNameAndAddress()
-    log.debug(`Blocked contact ${name} (id = ${contactId})`)
   }
 
   createChatByContactId (contactId) {
@@ -158,23 +100,6 @@ class DeltaChatController extends EventEmitter {
     return true
   }
 
-  deleteChat (chatId) {
-    log.debug(`action - deleting chat ${chatId}`)
-    this._dc.deleteChat(chatId)
-    this._render()
-  }
-
-  archiveChat (chatId, archive) {
-    log.debug(`action - archiving chat ${chatId}`)
-    this._dc.archiveChat(chatId, archive)
-    this._render()
-  }
-
-  showArchivedChats (show) {
-    this._showArchivedChats = show
-    this._render()
-  }
-
   createGroupChat (verified, name, image, contactIds) {
     let chatId
     if (verified) chatId = this._dc.createVerifiedGroupChat(name)
@@ -183,18 +108,6 @@ class DeltaChatController extends EventEmitter {
     contactIds.forEach(id => this._dc.addContactToChat(chatId, id))
     this.selectChat(chatId)
     return { chatId }
-  }
-
-  leaveGroup (chatId) {
-    log.debug(`action - leaving chat ${chatId}`)
-    this._dc.removeContactFromChat(chatId, C.DC_CONTACT_ID_SELF)
-  }
-
-  selectChat (chatId) {
-    log.debug(`action - selecting chat ${chatId}`)
-    this._pages = 1
-    this._selectedChatId = chatId
-    this._render()
   }
 
   /**
@@ -216,38 +129,6 @@ class DeltaChatController extends EventEmitter {
     return this._dc.getSecurejoinQrCode(chatId)
   }
 
-  checkPassword (password) {
-    return password === this.getConfig('mail_pw')
-  }
-
-  keysImport (directory) {
-    this._dc.importExport(C.DC_IMEX_IMPORT_SELF_KEYS, directory)
-  }
-
-  keysExport (directory) {
-    this._dc.importExport(C.DC_IMEX_EXPORT_SELF_KEYS, directory)
-  }
-
-  backupExport (dir) {
-    this._dc.importExport(C.DC_IMEX_EXPORT_BACKUP, dir)
-  }
-
-  setConfig (key, value) {
-    log.info(`Setting config ${key}:${value}`)
-    return this._dc.setConfig(key, String(value))
-  }
-
-  getConfig (key) {
-    return this._dc.getConfig(key)
-  }
-
-  getConfigFor (keys) {
-    let config = {}
-    for (let key of keys) {
-      config[key] = this.getConfig(key)
-    }
-    return config
-  }
 
   setLocation (latitude, longitude, accuracy) {
     return this._dc.setLocation(latitude, longitude, accuracy)
@@ -289,193 +170,7 @@ class DeltaChatController extends EventEmitter {
     return '#' + integerColor.toString(16)
   }
 
-  _chatList (showArchivedChats) {
-    if (!this._dc) return []
 
-    const listFlags = showArchivedChats ? C.DC_GCL_ARCHIVED_ONLY : 0
-    const list = this._dc.getChatList(listFlags, this._query)
-    const listCount = list.getCount()
-
-    const chatList = []
-    for (let i = 0; i < listCount; i++) {
-      const chatId = list.getChatId(i)
-      const chat = this._getChatById(chatId)
-
-      if (!chat) continue
-
-      if (chat.id === C.DC_CHAT_ID_DEADDROP) {
-        const messageId = list.getMessageId(i)
-        chat.deaddrop = this._deadDropMessage(messageId)
-      }
-
-      // This is NOT the Chat Oject, it's a smaller version for use as ChatListItem in the ChatList
-      chatList.push({
-        id: chat.id,
-        summary: list.getSummary(i).toJson(),
-        name: chat.name,
-        deaddrop: chat.deaddrop,
-        freshMessageCounter: chat.freshMessageCounter,
-        profileImage: chat.profileImage,
-        color: chat.color,
-        isVerified: chat.isVerified,
-        isGroup: chat.isGroup
-      })
-    }
-    return chatList
-  }
-
-  _getGeneralFreshMessageCounter () {
-    const list = this._dc.getChatList(0, this._query)
-    const listCount = list.getCount()
-
-    var freshMessageCounter = 0
-    for (let i = 0; i < listCount; i++) {
-      const chatId = list.getChatId(i)
-      const chat = this._dc.getChat(chatId).toJson()
-
-      if (!chat) continue
-
-      if (chat.id !== C.DC_CHAT_ID_DEADDROP) {
-        freshMessageCounter += this._dc.getFreshMessageCount(chatId)
-      }
-    }
-    return freshMessageCounter
-  }
-
-  _deadDropMessage (id) {
-    const msg = this._dc.getMessage(id)
-    const fromId = msg && msg.getFromId()
-
-    if (!fromId) {
-      log.warn('Ignoring DEADDROP due to missing fromId')
-      return
-    }
-
-    const contact = this._dc.getContact(fromId).toJson()
-    return { id, contact }
-  }
-
-  _selectedChat (showArchivedChats, chatList, selectedChatId) {
-    let selectedChat = this._getChatById(selectedChatId)
-    if (!selectedChat) return null
-    if (selectedChat.id !== C.DC_CHAT_ID_DEADDROP) {
-      if (selectedChat.freshMessageCounter > 0) {
-        this._dc.markNoticedChat(selectedChat.id)
-        selectedChat.freshMessageCounter = 0
-      }
-
-      if (this._saved.markRead) {
-        this._dc.markSeenMessages(selectedChat.messages.map((msg) => msg.id))
-      }
-    }
-
-    return selectedChat
-  }
-
-  _getChatById (chatId) {
-    if (!chatId) return null
-    const rawChat = this._dc.getChat(chatId)
-    if (!rawChat) return null
-
-    const chat = rawChat.toJson()
-    let draft = this._dc.getDraft(chatId)
-
-    if (draft) {
-      chat.draft = draft.getText()
-    } else {
-      chat.draft = ''
-    }
-
-    var messageIds = this._dc.getChatMessages(chat.id, C.DC_GCM_ADDDAYMARKER, 0)
-    // This object is NOT created with object assign to promote consistency and to be easier to understand
-    return {
-      id: chat.id,
-      name: chat.name,
-      isVerified: chat.isVerified,
-      profileImage: chat.profileImage,
-
-      archived: chat.archived,
-      subtitle: chat.subtitle,
-      type: chat.type,
-      isUnpromoted: chat.isUnpromoted,
-      isSelfTalk: chat.isSelfTalk,
-
-      contacts: this._dc.getChatContacts(chatId).map(id => this._dc.getContact(id).toJson()),
-      totalMessages: messageIds.length,
-      messages: this._messagesToRender(messageIds),
-      color: this._integerToHexColor(chat.color),
-      summary: undefined,
-      freshMessageCounter: this._dc.getFreshMessageCount(chatId),
-      isGroup: isGroupChat(chat),
-      isDeaddrop: chatId === C.DC_CHAT_ID_DEADDROP,
-      draft: chat.draft
-    }
-  }
-
-  _messagesToRender (messageIds) {
-    const countMessages = messageIds.length
-    const messageIdsToRender = messageIds.splice(
-      Math.max(countMessages - (this._pages * PAGE_SIZE), 0),
-      countMessages
-    )
-
-    if (messageIdsToRender.length === 0) return []
-
-    let messages = Array(messageIdsToRender.length)
-
-    for (let i = messageIdsToRender.length - 1; i >= 0; i--) {
-      let id = messageIdsToRender[i]
-      let json = this.messageIdToJson(id)
-
-      if (id === C.DC_MSG_ID_DAYMARKER) {
-        json.daymarker = {
-          timestamp: messages[i + 1].msg.timestamp,
-          id: 'd' + i
-        }
-      }
-      messages[i] = json
-    }
-
-    return messages
-  }
-
-  messageIdToJson (id) {
-    const msg = this._dc.getMessage(id)
-    const filemime = msg.getFilemime()
-    const filename = msg.getFilename()
-    const filesize = msg.getFilebytes()
-    const fromId = msg.getFromId()
-    const isMe = fromId === C.DC_CONTACT_ID_SELF
-    const setupCodeBegin = msg.getSetupcodebegin()
-    const contact = fromId ? this._dc.getContact(fromId).toJson() : {}
-    if (contact.color) {
-      contact.color = this._integerToHexColor(contact.color)
-    }
-
-    return {
-      id,
-      msg: msg.toJson(),
-      filemime,
-      filename,
-      filesize,
-      fromId,
-      isMe,
-      contact,
-      isInfo: msg.isInfo(),
-      setupCodeBegin
-    }
-  }
-
-  fetchMessages () {
-    this._pages++
-    this._render()
-  }
-
-  forwardMessage (msgId, contactId) {
-    const chatId = this._dc.getChatIdByContactId(contactId)
-    this._dc.forwardMessages(msgId, chatId)
-    this.selectChat(chatId)
-  }
 
   _blockedContacts () {
     if (!this._dc) return []
@@ -499,10 +194,12 @@ class DeltaChatController extends EventEmitter {
     return this._dc.getContactEncryptionInfo(contactId)
   }
 
-  getChatMedia (msgType1, msgType2) {
-    if (!this._selectedChatId) return
-    const mediaMessages = this._dc.getChatMedia(this._selectedChatId, msgType1, msgType2)
-    return mediaMessages.map(this.messageIdToJson.bind(this))
+
+  isGroupChat (chat) {
+    return [
+      C.DC_CHAT_TYPE_GROUP,
+      C.DC_CHAT_TYPE_VERIFIED_GROUP
+    ].includes(chat && chat.type)
   }
 
   /**
@@ -520,57 +217,5 @@ class DeltaChatController extends EventEmitter {
   }
 }
 
-function addServerFlags (credentials) {
-  return Object.assign({}, credentials, {
-    serverFlags: serverFlags(credentials)
-  })
-}
-
-function isGroupChat (chat) {
-  return [
-    C.DC_CHAT_TYPE_GROUP,
-    C.DC_CHAT_TYPE_VERIFIED_GROUP
-  ].includes(chat && chat.type)
-}
-
-function serverFlags ({ mailSecurity, sendSecurity }) {
-  const flags = []
-
-  if (mailSecurity === 'ssl') {
-    flags.push(C.DC_LP_IMAP_SOCKET_SSL)
-  } else if (mailSecurity === 'starttls') {
-    flags.push(C.DC_LP_IMAP_SOCKET_STARTTLS)
-  } else if (mailSecurity === 'plain') {
-    flags.push(C.DC_LP_SMTP_SOCKET_PLAIN)
-  }
-
-  if (sendSecurity === 'ssl') {
-    flags.push(C.DC_LP_SMTP_SOCKET_SSL)
-  } else if (sendSecurity === 'starttls') {
-    flags.push(C.DC_LP_SMTP_SOCKET_STARTTLS)
-  } else if (sendSecurity === 'plain') {
-    flags.push(C.DC_MAX_GET_INFO_LEN)
-  }
-
-  if (!flags.length) return null
-
-  return flags.reduce((flag, acc) => {
-    return acc | flag
-  }, 0)
-}
-
-if (!module.parent) {
-  // TODO move this to unit tests
-  console.log(serverFlags({
-    mailSecurity: 'ssl',
-    sendSecurity: 'ssl'
-  }))
-  console.log(C.DC_LP_IMAP_SOCKET_SSL | C.DC_LP_SMTP_SOCKET_SSL)
-  console.log(serverFlags({
-    mailSecurity: 'starttls',
-    sendSecurity: 'starttls'
-  }))
-  console.log(C.DC_LP_IMAP_SOCKET_STARTTLS | C.DC_LP_SMTP_SOCKET_STARTTLS)
-}
 
 module.exports = DeltaChatController
